Move ShareService tests to test/services
